Refactor delete chart dialog components structure
Moved delete-chart-buttons.tsx and delete-chart-dialog.tsx into a new delete-chart-dialog directory for better organization. Updated import paths accordingly in header-right-buttons.tsx.
