+++ components/hospital/icu/main/chart/selected-chart/chart-header/header-right-buttons/delete-chart-dialog/delete-chart-dialog.tsx
